+++ lib/vite/view.ex
@@ -2,7 +2,7 @@ defmodule Vite.View do
   @moduledoc """
   Help with View integration into Phoenix views
   """
-  alias Vite.{Config, Entry, ManifestReader}
+  alias Vite.{Config, ManifestReader}
 
   @doc """
   The snippet for `@vite/client` during development. Does nothing in :prod env.
@@ -47,18 +47,35 @@ defmodule Vite.View do
   <link rel="modulepreload" href="/assets/_vendor.7788aaa.js">
   `
   """
-  @spec for_entry(Entry.t(), binary()) :: binary()
-  def for_entry(entry = %Entry{}, prefix \\ "/") do
-    script = entry.file |> module_script(prefix)
-    imports = entry.imports |> Enum.map(&module_preload(&1, prefix)) |> Enum.join("\n")
-    css_files = entry.cssfiles |> Enum.map(&css_link(&1, prefix)) |> Enum.join("\n")
-    [css_files, script, imports] |> Enum.join("\n")
+  @spec for_entry(list(), binary()) :: binary()
+  def for_entry(entry, prefix \\ "/") do
+    entry |> Enum.map(fn file_tuple -> handle(file_tuple, prefix) end) |> Enum.join("\n")
+  end
+
+  defp handle({:css, file}, prefix) do
+    css_link(file, prefix)
+  end
+
+  defp handle({:import_css, file}, prefix) do
+    css_link(file, prefix)
+  end
+
+  defp handle({:module, file}, prefix) do
+    module_script(file, prefix)
+  end
+
+  defp handle({:import_module, file}, prefix) do
+    module_preload(file, prefix)
+  end
+
+  defp handle(_, _prefix) do
+    nil
   end
 
   @doc """
   Helper to get HTML for all entry points at once
   """
-  @spec for_entries(list(Entry.t())) :: binary()
+  @spec for_entries(list(list)) :: binary()
   def for_entries(entries, prefix \\ "/") do
     entries |> Enum.map(&for_entry(&1, prefix)) |> Enum.join("\n")
   end